+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 3.8.0
 
+* added `Search` to `Filter`
+* added `Sort` to `Filter` and exposed fields
 * added `Offset` to `Filter`
-* added `OrderBy` to `Filter` and exposed fields
 * updated referrer blacklist
 * updated referrer mapping
 * updated dependencies